fix(package): use the same Babel conf as elsewhere
This introduces the same Babel version and configuration as is used
elsewhere. Prior to this an older version of Babel with a deprecated
version of core-js was used for polyffiling. You know the version that
caused Vis Util and subsequently everything else not to work in IE 11
due to missing polyfills for some newer features.